Cleaned HomeController
